p40894: 13.71 seconds → 3.19 seconds → 3.15 seconds → 3.10 seconds

p15240 3.29 seconds → 0.843 seconds → 0.827 seconds